skincare-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.2
pymongo==4.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
opencv-python==4.8.1.78
numpy==1.26.4
pillow==10.1.0
httpx==0.25.0
pytest==7.4.3
pytest-asyncio==0.21.1
email-validator==2.1.0